% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GK2011-package.R
\docType{package}
\name{GK2011-package}
\alias{GK2011}
\alias{GK2011-package}
\title{GK2011}
\description{
Gaines and Kuklinski (2011) Estimators for Hybrid Experiments
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
Brian J. Gaines and James H. Kuklinski, (2011), "Experimental Estimation of Heterogeneous Treatment Effects Related to Self-Selection," \emph{American Journal of Political Science} 55(3): 724-736.
}
\seealso{
\code{\link{estimate}}
}
\keyword{package}

